# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Apache Airavata HTTP API
  version: "1.0.0"
  description: |
    The REST API served by `airavata-server` (see `internal/server/server.go`).

    All request and response bodies are JSON (`Content-Type: application/json`).

    ## Authentication

    Authentication is a bearer token validated by CILogon introspection, carried as
    `Authorization: Bearer <token>`. A request without a token is anonymous rather
    than rejected at the door: each operation decides for itself whether an anonymous
    caller is acceptable. A token that is present but invalid is always `401`, with
    `WWW-Authenticate: Bearer error="invalid_token"`. If the identity provider cannot
    be reached the call fails `502`, because that is not the caller's fault.

    ## Authorization

    Two schemes are in play, depending on the resource:

    - **Platform roles** (`ADMIN`, `SUPER_ADMIN`) guard the shared catalogue —
      Slurm clusters, partitions, application templates and batch deployments.
      Reads of the catalogue are open to anyone, including anonymous callers;
      writes require a role.
    - **Ownership and sharing** guards the per-user resources — SSH keys, groups,
      cluster configs, SCP data storages, data products and processes. Ownership
      comes from the token and is never carried in a request body, so nothing can be
      created on someone else's behalf or handed over by editing it. None of these
      are readable anonymously.

...

...

...

...
    state.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/users:
    get:
      tags: [Users]
      summary: List users
      description: Requires `ADMIN` or `SUPER_ADMIN` authority.
      operationId: listUsers
      responses:
        "200":
          description: Every registered user.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/UserResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Users]
      summary: Register a user
      description: |
        Creates the `users` row an owned resource is attached to. Owning anything in
        this API requires such a row.
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserRegistration" }
      responses:
        "200":
          description: The registered user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/users/{userId}:
    parameters:
      - $ref: "#/components/parameters/UserId"
    get:
      tags: [Users]
      summary: Read a user
      description: |
        The response carries no email, auth method or roles — widening it would leak
        more about an account than this API ever has.
      operationId: getUser
      responses:
        "200":
          description: The user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Users]
      summary: Update a user
      description: |
        Update the user with the provided information.
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UserRegistration" }
      responses:
        "200":
          description: The updated user.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }

  # --------------------------------------------------------------- Groups
...
  /api/v1/processes/{processId}/statuses:
    parameters:
      - $ref: "#/components/parameters/ProcessId"
    get:
      tags: [Processes]
      summary: List a process's status history
      description: |
        Read-only. Statuses are recorded internally — by process submission and by
        whatever submits and monitors the actual job — never accepted from a client,
        so there is no `POST` or `PUT` here.
      operationId: listProcessStatuses
      responses:
        "200":
          description: Every status recorded for the process.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/StatusResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/processes/{processId}/statuses/{statusId}:
    parameters:
      - $ref: "#/components/parameters/ProcessId"
      - name: statusId
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Processes]
      summary: Read one recorded status
      operationId: getProcessStatus
      responses:
        "200":
          description: The status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StatusResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "502": { $ref: "#/components/responses/BadGateway" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A CILogon access token, validated by RFC 7662 introspection. The bootstrap
        root token is also accepted; see INSTALL.md.
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string }
...
    ProcessId:
      name: processId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    BadRequest:
      description: |
        Malformed JSON, or a body that failed its constraints. Validation failures
        carry `fieldErrors`.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          examples:
            validation:
              summary: Validation failed
              value:
                status: 400
                error: Bad Request
                message: Validation failed
                fieldErrors:
                  - field: clusterName
                    message: Cluster name cannot be blank
    Unauthorized:
      description: |
        No token where one is required, or a token that is present but unusable. An
        invalid token also carries `WWW-Authenticate: Bearer error="invalid_token"`.
      headers:
        WWW-Authenticate:
          description: Present when the token was rejected.
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated, but lacking the authority for this call.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        No such record — or, for a group, a record the caller has no standing to know
        exists.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            status: 404
            error: Not Found
            message: "Slurm cluster not found: c1d2e3f4-5678-4abc-9def-0123456789ab"
...
    BadGateway:
      description: |
        The identity provider could not be reached to validate the token. Not the
        caller's fault, so it is not reported as an auth failure.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InternalError:
      description: |
        An unexpected server failure. The message is always `"Internal server error"`;
        the detail is logged rather than returned, since it may name internal state.

        A `502` is also possible on any authenticated call if the identity provider is
        unreachable.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
...
    FieldError:
      type: object
      description: One failed validation constraint.
      properties:
        field:
          type: string
          description: |
            The JSON path of the offending field, including indexes and nesting.
          examples:
            ["inputs[0].inputType", "defaultBatchJobConfig.wallTimeMinutes"]
        message: { type: string }
      required: [field, message]
    Error:
      type: object
      description: The envelope every failure response uses.
      properties:
        status:
          type: integer
          description: The HTTP status, repeated in the body.
        error:
          type: string
          description: The standard reason phrase for `status`.
        message:
          type: string
          description: Human-readable detail.
        fieldErrors:
          type: array
          description: |
            Present only on request-body validation failures, listing every
            constraint that failed so a client can report all of them at once.
          items: { $ref: "#/components/schemas/FieldError" }
      required: [status, error, message]

    # ----------------------------------------------------------- Enums
    AuthMethod:
      type: string
      enum: [CILOGON, SYSTEM]
    UserStatus:
      type: string
      enum: [ACTIVE, INACTIVE, SUSPENDED]
...
    ProcessStatusType:
      type: string
      enum: [CREATED, SUBMITTED, RUNNING, COMPLETED, FAILED]
...
    UserRegistration:
      type: object
      description: The create/update payload for a user.
      properties:
        userId: { type: string }
        email: { type: [string, "null"], format: email }
        firstName: { type: [string, "null"] }
        lastName: { type: [string, "null"] }
        authMethod:
          oneOf:
            - { $ref: "#/components/schemas/AuthMethod" }
            - { type: "null" }
      required: [userId]
    UserResponse:
      type: object
      description: |
        The read model for a user. It deliberately omits email, auth method and roles.
      properties:
        userId: { type: string }
        status:
          oneOf:
            - { $ref: "#/components/schemas/UserStatus" }
            - { type: "null" }
        firstName: { type: [string, "null"] }
        lastName: { type: [string, "null"] }
        createdAt:
          type: integer
          format: int64
          description: Milliseconds since the epoch.
      required: [userId, status, firstName, lastName, createdAt]
...
    StatusResponse:
      type: object
      description: |
        One recorded process status. There is no request counterpart: statuses are
        never created or updated from a client body.
      properties:
        processStatusId: { type: string, format: uuid }
        processId: { type: [string, "null"], format: uuid }
        status:
          oneOf:
            - { $ref: "#/components/schemas/ProcessStatusType" }
            - { type: "null" }
        log: { type: [string, "null"] }
        timestamp:
          type: [integer, "null"]
          format: int64
          description: When the status was recorded, in milliseconds since the epoch.
      required: [processStatusId]
...